Module itron_solid_fmp3::error

This is supported on crate feature unstable only.

Error types.

Errors indicate exceptional outcomes of system calls. Like most traditional operating systems, they are represented by error codes in μITRON.

Error Kind Types

TODO: describe error kind types

Critical Errors

The following errors may be escalated to panics or undefined behaviors:

Kernel integrity errors (some cases of E_SYS in the TOPPERS kernels).

Memory access permission errors (E_MACV) caused by pointers that are supposed to be accessible by the current thread (e.g., pointers referring to local variables).

E_PAR caused by invalid timeout values.

Structs

Error

Target-specific error value that can be categorized as one of the error kinds represented by Kind.

ErrorCode

Raw error code.

Kind

Type for error kinds that are valid in the current target kernel.

Enums

Never

Indicates that the error kind never occurs in the current target kernel.

Traits

ErrorKind

Trait for error kind types.

Type Definitions

MaybeKind

Placeholder for error kind variants.