Struct itron_solid_fmp3::mutex::Mutex

pub struct Mutex(_);

This is supported on crate features unstable and dcre only.

An owned mutex.

Deletes the mutex automatically when dropped. The destructor will panic if the deletion fails.

Implementations

impl Mutex

pub fn build() -> Builder

acre_mtx: Create a builder for Mutex.

Examples

use itron::mutex::Mutex;
let mutex = Mutex::build()
    .finish()
    .expect("failed to create a mutex");

mutex.as_ref().lock()
   .expect("failed to perform a lock operation");
mutex.as_ref().unlock()
   .expect("failed to perform a unlock operation");

use itron::mutex::{Mutex, PriorityProtection};
let mutex = Mutex::build()
    .priority_protection(PriorityProtection::Ceiling { priority: 4 })
    .finish()
    .expect("failed to create a priority-ceiling mutex");

mutex.as_ref().lock()
   .expect("failed to perform a lock operation");
mutex.as_ref().unlock()
   .expect("failed to perform a unlock operation");

impl Mutex

pub const unsafe fn from_raw_nonnull(id: NonNullID) -> Self

Construct a Mutex from a raw object ID.

Safety

See Object ID Wrappers.

pub const fn leak<'a>(self) -> MutexRef<'a>

Consume and “leak” self, returning a reference MutexRef<'a>.

pub const fn as_raw(&self) -> ID

Get the raw object ID.

pub const fn as_raw_nonnull(&self) -> NonNullID

Get the raw object ID as abi::NonNullID.

pub const fn as_ref(&self) -> MutexRef<'_>

Borrow Mutex as MutexRef.

Use this to perform operations on mutexes because most of the methods are implemented on MutexRef but not Mutex.