Module itron_solid_fmp3::task

This is supported on crate feature unstable only.

Tasks

Structs

AllProcessors

An instance of IntoProcessorSet specifying all processors.

BadContextError

The error type returned by current when the CPU lock state is active, or the current thread is not in a task context.

Builderdcre

The builder type for tasks. Created by Task::build.

Current

Represents a reference to the current task. Returned by current.

Info

Task information returned by TaskRef::info.

Taskdcre

An owned task.

TaskRef

A borrowed reference to a task.

Enums

ActivateError

Error type for TaskRef::activate.

ActivateOnError

Error type for TaskRef::activate_on.

BuildErrordcre

Error type for Task::build.

CancelActivateAllError

Error type for TaskRef::cancel_activate_all.

CancelWakeAllError

Error type for TaskRef::cancel_wake_all.

CurrentIdError

Error type for current_id.

DelayError

Error type for delay.

DeleteErrordcre

Error type for TaskRef::delete.

DisableTerminationError

Error type for disable_termination.

ExitError

Error type for exit.

InfoError

Error type for TaskRef::info.

MigrateError

Error type for TaskRef::migrate.

PriorityError

Error type for TaskRef::priority.

RaiseTerminationError

Error type for TaskRef::raise_termination.

ReleaseWaitError

Error type for TaskRef::release_wait.

ResumeError

Error type for TaskRef::resume.

SetBasePriorityError

Error type for TaskRef::set_base_priority.

SleepError

Error type for sleep.

SleepTimeoutError

Error type for sleep_timeout.

State

Task state returned by TaskRef::state.

StateError

Error type for TaskRef::state.

SuspendError

Error type for TaskRef::suspend.

TerminateError

Error type for TaskRef::terminate.

WakeError

Error type for TaskRef::wake.

Traits

IntoProcessorSetdcre

The trait implemented by types that can be passed to crate::task::Builder::processor_affinity. This trait is sealed.

Functions

current

Get a reference to the current task.

current_id

get_tid: Get the currently running task’s ID.

delay

dly_tsk: Delay the current task.

disable_termination

dis_ter: Disable the termination of the current task by a termination request.

enable_termination

ena_ter: Re-enable the termination of the current task by a termination request.

exit⚠

ext_tsk: Terminate the current task.

is_termination_disabled

sns_ter: Determine if the termination by a termination request is disabled for the current task.

sleep

slp_tsk: Put the current task to sleep.

sleep_timeout

dly_tsk: Put the current task to sleep with timeout.

Type Definitions

EnableTerminationError

Error type for enable_termination.

Priority

Task priority value.