Struct itron_solid_fmp3::task::TaskRef

pub struct TaskRef<'a> { /* fields omitted */ }

This is supported on crate feature unstable only.

A borrowed reference to a task.

Implementations

impl TaskRef<'_>

Object ID conversion

pub const unsafe fn from_raw_nonnull(id: NonNullID) -> Self

Construct a TaskRef from a raw object ID.

Safety

See Object ID Wrappers.

pub const fn as_raw(self) -> ID

Get the raw object ID.

pub const fn as_raw_nonnull(self) -> NonNullID

Get the raw object ID as abi::NonNullID.

impl TaskRef<'_>

Management

pub fn activate(self) -> Result<(), Error<ActivateError>>

act_tsk: Pend an activation request for the task.

pub fn activate_on(
self,
processor: Processor
) -> Result<(), Error<ActivateOnError>>

mact_tsk: Pend an activation request for the task, assigning it to the specified processor.

pub fn cancel_activate_all(self) -> Result<usize, Error<CancelActivateAllError>>

can_act: Cancel any pending activation requests for the task. Returns the number of cancelled requests.

pub fn set_base_priority(
self,
new_priority: Priority
) -> Result<(), Error<SetBasePriorityError>>

chg_pri: Change the task’s base priority.

pub fn priority(self) -> Result<Priority, Error<PriorityError>>

get_pri: Get the task’s priority.

Use TaskRef::info and Info::base_priority to get the base priority.

pub fn migrate(self, processor: Processor) -> Result<(), Error<MigrateError>>

mig_tsk: Change the task’s assigned processor.

pub unsafe fn delete(self) -> Result<(), Error<DeleteError>>

This is supported on crate feature dcre only.

del_tsk: Delete the task.

Safety

See Object ID Wrappers.

pub fn state(self) -> Result<State, Error<StateError>>

get_tst: Get the task’s state.

pub fn info(self) -> Result<Info, Error<InfoError>>

ref_tsk: Get the task’s general information.

impl TaskRef<'_>

Synchronization

pub fn wake(self) -> Result<(), Error<WakeError>>

wup_tsk: Pend a wake up request for the task.

This method and the sleep function are semantically analogous to std::thread::Thread::unpark and std::thread::park, respectively. However, unlike unpark, this method will return Err(WakeError::QueueOverflow) if the token is already present.

pub fn cancel_wake_all(self) -> Result<usize, Error<CancelWakeAllError>>

can_wup: Cancel any wake up requests for the task. Returns the number of cancelled requests.

pub fn release_wait(self) -> Result<(), Error<ReleaseWaitError>>

rel_wai: Forcibly release the task from waiting.

pub fn suspend(self) -> Result<(), Error<SuspendError>>

sus_tsk: Suspend the task.

pub fn resume(self) -> Result<(), Error<ResumeError>>

rsm_tsk: Resume the task.

impl TaskRef<'_>

Termination

pub unsafe fn terminate(self) -> Result<(), Error<TerminateError>>

ter_tsk: Terminate the task.

Safety

If the task’s stack is reused later, stored local variables are destroyed without running their destructors, violating the pinning requirements.

pub unsafe fn raise_termination(
self
) -> Result<(), Error<RaiseTerminationError>>

ras_ter: Pend a termination request.

Safety

If the task’s stack is reused later, stored local variables are destroyed without running their destructors, violating the pinning requirements.